% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca_plot.R
\name{lca_plot}
\alias{lca_plot}
\title{LCA-Plot}
\usage{
lca_plot(x)
}
\arguments{
\item{x}{the fitted object created by poLCA::poLCA()}
}
\value{
the LCA-plot
}
\description{
This function creates a plot from an estimated latent class solution created with the package {poLCA}. It uses the "probs"-Variables of the object.
}
